1) improving fairness for balance scheduling, but not complete fairness cannot be guaranteed because load balancing period is too long to keep fine-grained fairness. Can observe a situation in which large imbalance by balance scheduling retains before load balancing is triggered. It is made worse since load balance interval gets longer for busy cores. TODO: enhance fairness more
2) interactive_count variable is added into per-cpu variable and
task_group struct. The per-cpu interactive count will be used for
isolating other vcpus hosting batch workloads from a core hosting
interactive workloads. And an interactive count in task_group struct is
used for considering threads in a VM group other than vcpus, e.g., QEMU
main and display threads. When the interactive count is positive, other
threads in this group are considered to be interactive as well.

3) /proc/sys/kernel/ parameters are added to control some tweaks of the
CFS scheduler.
 - sched_interactive_preempt: to control the preemption decision
 - sched_aggressive_load: to truncate load_avg and load_period when a
   cfs rq has no load at all. This increases a share for each vcpu at
   any given time if vcpus do not run at the same time, i.e.,
   thread-level parallelism is closed to one.