1) bug fix for balance scheduling: At booting a VM, vcpu threads are recognized as vcpus with se.is_vcpu, the QEMU main thread temporarily replaces them, and finally the vcpu threads are set as vcpus again. But, while the QEMU main thread is regarded as a vcpu, it creates some number of threads such as a spice rendering thread. This created thread is wrongly classified as a vcpu. So, cfs_rq->nr_running_vcpus doesn't work well. Fixed by assigning NEW_VCPU_SE to se.is_vcpu when each vcpu is created with no further updates.
2) Fairness enhancement for balance scheduling: Add
cause_load_imbalance() function, which returns whether a thread causes
load imbalance if it is woken up at a cpu. If load imbalance is
possible, balance scheduling cannot place a thread on the cpu. The
extent to which the balance scheduling guarantees fairness is expressed
as 'balsched_load_imbalance_pct', which is 125 by default. From the
experiment of four parsec's bodytracks as normal threads,
lb_imbalance / lb_count is almost the same as the baseline case (i.e.,
neither affinity nor balance)

3) CPU lag monitoring: So experimental and being developed. But not sure if
this feature is necessary, so pending at this time.